Add simple JDBC examples
There are two ways to run this:
- through JdbcExample.main(), which connects to the database manually
- through jdbcExample.jsp, which gets a connection from Tomcat

If running main(), mysql-connector-java needs to be on the classpath.

If running jdbcExample.jsp, mysql-connector-java needs to be installed
to Tomcat's lib directory.

mysql-connector-java can be downloaded like:

  mvn dependency:get -DgroupId=mysql -DartifactId=mysql-connector-java -Dversion=8.0.24

and then installed to Tomcat like:

  ln -nsf ../../.m2/repository/mysql/mysql-connector-java/8.0.24/mysql-connector-java-8.0.24.jar -t ~/apache-tomcat-7.0.109/lib/

The Tomcat DataSource can be configured with something like the
following in context.xml:

  <Resource
    name="jdbc/coffee"
    auth="Container"
    type="javax.sql.DataSource"
    driverClassName="com.mysql.jdbc.Driver"
    url="jdbc:mysql://localhost/coffee"
    username="user"
    password="password"
  />